remove app-wrapper, flat the structure
